M2 point 0: fix MS_ParseDate int*/long* mismatch (live bug, captions.c mark-by-date used uninitialized long locals); widen gtm to long* matching gtime()'s real width and every other existing caller, keep year/month/day/hour/min/sec/wday int*
FossilOrigin-Name: f4a9d6909b254cdb3599e8b1b30182b8702e7cfc7a7541104def69fd535e498a

+++ src/atkams/messages/lib/ams.c
@@ -187,7 +187,7 @@ struct ams *self;
     return(0);
 }
 
-long ams__CUI_DisambiguateDir(self, shortname, longname) 
+long ams__CUI_DisambiguateDir(self, shortname, longname)
 struct ams *self;
 char *shortname, *longname;
 {
@@ -424,7 +424,7 @@ char *tolist;
     return(0);
 }
 
-long ams__CUI_RewriteHeaderLine(self, addr, newaddr) 
+long ams__CUI_RewriteHeaderLine(self, addr, newaddr)
 struct ams *self;
 char *addr, *newaddr;
 {
@@ -656,13 +656,13 @@ char *filename;
 long ams__MS_ParseDate(self, indate, year, month, day, hour, min, sec, wday, gtm)
 struct ams *self;
 char *indate;
-long *year;
-long *month;
-long *day;
-long *hour;
-long *min;
-long *sec;
-long *wday;
+int *year;
+int *month;
+int *day;
+int *hour;
+int *min;
+int *sec;
+int *wday;
 long *gtm;
 {
     ReportMissing("MS_ParseDate");
