+++ src/atkams/messages/lib/amsn.c
@@ -112,7 +112,7 @@ struct amsn *self;
     return(CUI_DirectoriesToPurge());
 }
 
-long amsn__CUI_DisambiguateDir(self, shortname, longname) 
+long amsn__CUI_DisambiguateDir(self, shortname, longname)
 struct amsn *self;
 char *shortname, *longname;
 {
@@ -328,7 +328,7 @@ char *tolist;
     return(CUI_ResendMessage(cuid, tolist));
 }
 
-long amsn__CUI_RewriteHeaderLine(self, addr, newaddr) 
+long amsn__CUI_RewriteHeaderLine(self, addr, newaddr)
 struct amsn *self;
 char *addr, *newaddr;
 {
@@ -534,13 +534,13 @@ char *filename;
 long amsn__MS_ParseDate(self, indate, year, month, day, hour, min, sec, wday, gtm)
 struct amsn *self;
 char *indate;
-long *year;
-long *month;
-long *day;
-long *hour;
-long *min;
-long *sec;
-long *wday;
+int *year;
+int *month;
+int *day;
+int *hour;
+int *min;
+int *sec;
+int *wday;
 long *gtm;
 {
     return(MS_ParseDate(indate, year, month, day, hour, min, sec, wday, gtm));
